Don't pollute context menu
The context menu could be used anywhere, not only for "historable" files and it is not necessary to show menu items of `local-history` everywhere. For example "Show history of current file" is useless in context menu for toolbar. So this PR fix it -> it will show menu items only for workspace context menu.